% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{baby_name_maxdiff}
\alias{baby_name_maxdiff}
\title{Start the Baby Name MaxDiff Shiny app}
\usage{
baby_name_maxdiff()
}
\description{
Will start a local Shiny app that will walk the user through a MaxDiff exercise
}
